refactor: use-menu-button を Menu モジュールへ移す
消費者は components/Menu/index.ts の再エクスポート経由だけで、Menu 機能
の外から使われていない。credentials / overlay / siteProfile が use-*.ts
を機能ディレクトリに同居させているのと同じ形に揃える。

これで src/hooks/ は空になるため消える。

構造ドキュメントも実態に合わせる。hooks/useMenuButton.ts と
utils/get-profile-error-message.ts / get-registry-keys.ts は、いずれも
現存しないファイルを指していた。

ref #219
