+++ apps/inspector/src/components/Menu/index.ts
@@ -1,4 +1,4 @@
-export { useMenuButton } from "../../hooks/use-menu-button";
 export { Menu } from "./Menu";
 export { MenuButton } from "./MenuButton";
 export { MenuItem } from "./MenuItem";
+export { useMenuButton } from "./use-menu-button";
