+++ apps/lens/src/components/Menu/use-menu-button.ts
